requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
dnspython==2.4.2
